requirements.txt
pytest==3.9.1
python-telegram-bot==11.1.0
requests-mock==1.5.2
beautifulsoup4==4.6
idna==2.7
mpymysql==0.9.2